+++ RELEASE-NOTES.md
@@ -27,5 +27,6 @@
 
 ## 0.0.5
 
+- Add ability to display placeholder row inside JUPTable component when no row data is available (no transactions, no assets, etc...)
 - Upgrade JupTable sorting mechanisms; with room for extensions to more complex object types such as BigNumber and Date
 - Adds current page indicator to navigation drawer